refactor vue tests
